refactor: move the settings and DNS-account endpoints out of the closure
Seventh group of #669. Settings, DNSProviders, DNSAccounts and
DNSAccountDetail move into modules/api/resources_settings.py.

Closure complexity 335 -> 282; resources.py 2618 -> 2295 lines.

The four are grouped by what they configure rather than by how they are
registered, because those two do not line up here: Settings and DNSProviders
are registered by factory.py from the mapping create_api_resources returns,
while DNSAccounts and DNSAccountDetail are bound to a namespace inside that
function. Splitting a coherent module along that seam would let a registration
detail dictate the shape of the code, so both are handled explicitly at the
call site instead, and the returned mapping stays exactly what it was.

CAProviderTest sits beside these in the closure and is deliberately left for
its own group: at 273 lines it is a third of this block, and testing ACME CA
connectivity is not settings management.
